Major revision
Changed map_pkg to first generate a yaml file with all the config of the
map. Then spawns the elements of the map using Python3 nodes.
Modified the hexagon.sdf so that the vertexes are not moved by a scale
factor, but they are computed. This also avoids the width of the walls
to change.
Fixed the launch files. Still missing the pursuer-evader one :.(